+++ gui/backend/ios/native_platform.go
@@ -84,6 +84,8 @@ func (n *nativePlatform) SpellCheck(text string) []gui.SpellRange { return spell
 
 func (n *nativePlatform) SetWindowVibrancy(_ gui.VibrancyMaterial) {}
 
+func (n *nativePlatform) SetWindowOpacity(_ float32) {}
+
 // No window manager to hand a move or resize gesture to.
 func (n *nativePlatform) StartWindowDrag()                   {}
 func (n *nativePlatform) StartWindowResize(_ gui.WindowEdge) {}