+++ gui/native_platform_noop.go
@@ -45,6 +45,7 @@ func (noopNativePlatform) IMESetRect(_, _, _, _ int32)                         {
 func (noopNativePlatform) OpenURI(_ string) error                              { return nil }
 func (noopNativePlatform) TitlebarDark(_ bool)                                 {}
 func (noopNativePlatform) SetWindowVibrancy(_ VibrancyMaterial)                {}
+func (noopNativePlatform) SetWindowOpacity(_ float32)                          {}
 func (noopNativePlatform) StartWindowDrag()                                    {}
 func (noopNativePlatform) StartWindowResize(_ WindowEdge)                      {}
 func (noopNativePlatform) SpellCheck(_ string) []SpellRange                    { return nil }